Fix Linting issues (rename module)
