+++ recommender-back/src/services/data_fetcher.py
